+++ docs/changelog.rst
@@ -13,6 +13,7 @@ CHANGELOG
 - Add Apidae parser for Services
 - Add OpenStreetMap parser for infrastructures
 - Add OpenStreetMap parser for touristic contents
+- Add OpenStreetMap parser for Outdoor Sites
 
 **Bug fixes**